+++ src/main/java/dev/vality/codegen/generators/v3/ValityErlangServerGenerator.java
@@ -1,9 +1,10 @@
-package dev.vality.codegen.generators;
+package dev.vality.codegen.generators.v3;
 
-import io.swagger.codegen.*;
+import io.swagger.codegen.v3.*;
+import io.swagger.codegen.v3.generators.DefaultCodegenConfig;
 
-import io.swagger.models.Swagger;
-import io.swagger.util.Json;
+import io.swagger.v3.core.util.Json;
+import io.swagger.v3.oas.models.OpenAPI;
 
 import java.util.*;
 
@@ -16,12 +17,12 @@ import org.slf4j.LoggerFactory;
 
 import dev.vality.codegen.utils.erlang.*;
 
-public class ValityErlangServerGenerator extends DefaultCodegen implements CodegenConfig {
+public class ValityErlangServerGenerator extends DefaultCodegenConfig {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ValityErlangServerGenerator.class);
 
     protected String sourceFolder = "src";
-    protected String packageName = "swagger";
+    protected String packageName = "openapi";
     protected String apiPath = "src";
     protected String apiVersion = "1.0.0";
 
@@ -32,70 +33,71 @@ public class ValityErlangServerGenerator extends DefaultCodegen implements Codeg
         outputFolder = "generated-code/vality-erlang-server";
 
         /**
-         * Models.  You can write model files using the modelTemplateFiles map.
+         * Models. You can write model files using the modelTemplateFiles map.
          * if you want to create one template for file, you can do so here.
-         * for multiple files for model, just put another entry in the `modelTemplateFiles` with
+         * for multiple files for model, just put another entry in the
+         * `modelTemplateFiles` with
          * a different extension
          */
         modelTemplateFiles.clear();
 
         /**
-         * Api classes.  You can write classes for each Api file with the apiTemplateFiles map.
-         * as with models, add multiple entries with different extensions for multiple files per
+         * Api classes. You can write classes for each Api file with the
+         * apiTemplateFiles map.
+         * as with models, add multiple entries with different extensions for multiple
+         * files per
          * class
          */
         apiTemplateFiles.put(
-                "handler.mustache",   // the template to use
-                ".erl");       // the extension for each file to write
+                "handler.mustache", // the template to use
+                ".erl"); // the extension for each file to write
 
         /**
-         * Template Location.  This is the location which templates will be read from.  The generator
+         * Template Location. This is the location which templates will be read from.
+         * The generator
          * will use the resource stream to attempt to read the templates.
          */
         embeddedTemplateDir = templateDir = "vality-erlang-server";
 
         /**
-         * Reserved words.  Override this with reserved words specific to your language
+         * Reserved words. Override this with reserved words specific to your language
          */
         setReservedWordsLowerCase(
                 Arrays.asList(
                         "after", "and", "andalso", "band", "begin", "bnot", "bor", "bsl", "bsr", "bxor", "case",
                         "catch", "cond", "div", "end", "fun", "if", "let", "not", "of", "or", "orelse", "receive",
-                        "rem", "try", "when", "xor"
-                )
-        );
+                        "rem", "try", "when", "xor"));
 
         instantiationTypes.clear();
 
         typeMapping.clear();
 
-        typeMapping.put("enum",         "binary");
-        typeMapping.put("date",         "binary");
-        typeMapping.put("DateTime",     "binary");
-        typeMapping.put("string",       "binary");
-        typeMapping.put("char",         "binary");
-        typeMapping.put("binary",       "binary");
-        typeMapping.put("UUID",         "binary");
-        typeMapping.put("password",     "binary");
-        typeMapping.put("boolean",      "boolean");
-        typeMapping.put("integer",      "integer");
-        typeMapping.put("long",         "integer");
-        typeMapping.put("float",        "float");
-        typeMapping.put("double",       "float");
-        typeMapping.put("number",       "float");
-        typeMapping.put("array",        "list");
-        typeMapping.put("List",         "list");
-        typeMapping.put("map",          "map");
-        typeMapping.put("object",       "object");
-        typeMapping.put("file",         "file");
-        typeMapping.put("ByteArray",    "byte");
-
+        typeMapping.put("enum", "binary");
+        typeMapping.put("date", "binary");
+        typeMapping.put("DateTime", "binary");
+        typeMapping.put("string", "binary");
+        typeMapping.put("char", "binary");
+        typeMapping.put("binary", "binary");
+        typeMapping.put("UUID", "binary");
+        typeMapping.put("password", "binary");
+        typeMapping.put("boolean", "boolean");
+        typeMapping.put("integer", "integer");
+        typeMapping.put("long", "integer");
+        typeMapping.put("float", "float");
+        typeMapping.put("double", "float");
+        typeMapping.put("number", "float");
+        typeMapping.put("array", "list");
+        typeMapping.put("List", "list");
+        typeMapping.put("map", "map");
+        typeMapping.put("object", "object");
+        typeMapping.put("file", "file");
+        typeMapping.put("ByteArray", "byte");
 
         cliOptions.clear();
         cliOptions.add(new CliOption(CodegenConstants.PACKAGE_NAME, "Erlang package name (convention: lowercase).")
                 .defaultValue(this.packageName));
         /**
-         * Additional Properties.  These values can be passed to the templates and
+         * Additional Properties. These values can be passed to the templates and
          * are available in models, apis, and supporting files
          */
         additionalProperties.put("apiVersion", apiVersion);
@@ -112,22 +114,27 @@ public class ValityErlangServerGenerator extends DefaultCodegen implements Codeg
         }
 
         /**
-         * Supporting Files.  You can write single files for the generator with the
-         * entire object tree available.  If the input file has a suffix of `.mustache
-         * it will be processed by the template engine.  Otherwise, it will be copied
+         * Supporting Files. You can write single files for the generator with the
+         * entire object tree available. If the input file has a suffix of `.mustache
+         * it will be processed by the template engine. Otherwise, it will be copied
          */
         supportingFiles.add(new SupportingFile("rebar.config.mustache", "", "rebar.config"));
-        supportingFiles.add(new SupportingFile("app.src.mustache", "", "src" + File.separator + this.packageName + ".app.src"));
+        supportingFiles.add(
+                new SupportingFile("app.src.mustache", "", "src" + File.separator + this.packageName + ".app.src"));
         supportingFiles.add(new SupportingFile("router.mustache", "", toSourceFilePath("router", "erl")));
         supportingFiles.add(new SupportingFile("utils.mustache", "", toSourceFilePath("utils", "erl")));
         supportingFiles.add(new SupportingFile("types.mustache", "", toPackageNameSrcFile("erl")));
         supportingFiles.add(new SupportingFile("handler_api.mustache", "", toSourceFilePath("handler_api", "erl")));
         supportingFiles.add(new SupportingFile("logic_handler.mustache", "", toSourceFilePath("logic_handler", "erl")));
         supportingFiles.add(new SupportingFile("validation.mustache", "", toSourceFilePath("validation", "erl")));
-        supportingFiles.add(new SupportingFile("common_validator.mustache", "", toSourceFilePath("common_validator", "erl")));
-        supportingFiles.add(new SupportingFile("param_validator.mustache", "", toSourceFilePath("param_validator", "erl")));
-        supportingFiles.add(new SupportingFile("schema_validator.mustache", "", toSourceFilePath("schema_validator", "erl")));
-        supportingFiles.add(new SupportingFile("custom_validator.mustache", "", toSourceFilePath("custom_validator", "erl")));
+        supportingFiles
+                .add(new SupportingFile("common_validator.mustache", "", toSourceFilePath("common_validator", "erl")));
+        supportingFiles
+                .add(new SupportingFile("param_validator.mustache", "", toSourceFilePath("param_validator", "erl")));
+        supportingFiles
+                .add(new SupportingFile("schema_validator.mustache", "", toSourceFilePath("schema_validator", "erl")));
+        supportingFiles
+                .add(new SupportingFile("custom_validator.mustache", "", toSourceFilePath("custom_validator", "erl")));
         supportingFiles.add(new SupportingFile("schema.mustache", "", toSourceFilePath("schema", "erl")));
         writeOptional(outputFolder, new SupportingFile("README.mustache", "", "README.md"));
     }
@@ -149,7 +156,8 @@ public class ValityErlangServerGenerator extends DefaultCodegen implements Codeg
     }
 
     /**
-     * Configures a friendly name for the generator.  This will be used by the generator
+     * Configures a friendly name for the generator. This will be used by the
+     * generator
      * to select the library with the -l flag.
      *
      * @return the friendly name for the generator
@@ -160,7 +168,7 @@ public class ValityErlangServerGenerator extends DefaultCodegen implements Codeg
     }
 
     /**
-     * Returns human-friendly help for the generator.  Provide the consumer with help
+     * Returns human-friendly help for the generator. Provide the consumer with help
      * tips, parameters here
      *
      * @return A string value for the help message
@@ -180,8 +188,10 @@ public class ValityErlangServerGenerator extends DefaultCodegen implements Codeg
     }
 
     /**
-     * Escapes a reserved word as defined in the `reservedWords` array. Handle escaping
-     * those terms here.  This logic is only called if a variable matches the reserved words
+     * Escapes a reserved word as defined in the `reservedWords` array. Handle
+     * escaping
+     * those terms here. This logic is only called if a variable matches the
+     * reserved words
      *
      * @return the escaped term
      */
@@ -203,13 +213,14 @@ public class ValityErlangServerGenerator extends DefaultCodegen implements Codeg
     @Override
     public String addRegularExpressionDelimiter(String pattern) {
         if (pattern != null) {
-            return pattern.replaceAll("^/","").replaceAll("/$","");
+            return pattern.replaceAll("^/", "").replaceAll("/$", "");
         }
         return pattern;
     }
 
     /**
-     * Location to write api files.  You can use the apiPackage() as defined when the class is
+     * Location to write api files. You can use the apiPackage() as defined when the
+     * class is
      * instantiated
      */
     @Override
@@ -226,7 +237,8 @@ public class ValityErlangServerGenerator extends DefaultCodegen implements Codeg
     public String toOperationId(String operationId) {
         // method name cannot use reserved keyword, e.g. return
         if (isReservedWord(operationId)) {
-            LOGGER.warn(operationId + " (reserved word) cannot be used as method name. Renamed to " + camelize(sanitizeName("call_" + operationId)));
+            LOGGER.warn(operationId + " (reserved word) cannot be used as method name. Renamed to "
+                    + camelize(sanitizeName("call_" + operationId)));
             operationId = "call_" + operationId;
         }
 
@@ -252,10 +264,11 @@ public class ValityErlangServerGenerator extends DefaultCodegen implements Codeg
 
     @Override
     public Map<String, Object> postProcessSupportingFileData(Map<String, Object> objs) {
-        Swagger swagger = (Swagger) objs.get("swagger");
-        if (swagger != null) {
+        OpenAPI openAPI = (OpenAPI) objs.get("openapi");
+        if (openAPI != null) {
             try {
-                objs.put("swagger-json", Json.mapper().writer(new ErlangJsonPrinter()).with(new ErlangJsonFactory()).writeValueAsString(swagger));
+                objs.put("openapi-json", Json.mapper().writer(new ErlangJsonPrinter()).with(new ErlangJsonFactory())
+                        .writeValueAsString(openAPI));
             } catch (JsonProcessingException e) {
                 LOGGER.error(e.getMessage(), e);
             }
@@ -297,6 +310,11 @@ public class ValityErlangServerGenerator extends DefaultCodegen implements Codeg
     public String escapeUnsafeCharacters(String input) {
         // ref: http://stackoverflow.com/a/30421295/677735
         return input.replace("-ifdef", "- if def").replace("-endif", "- end if");
+    }
+
+    @Override
+    public String getDefaultTemplateDir() {
+        return templateDir;
     }
 
 }
