--- src/main/resources/META-INF/services/io.swagger.codegen.CodegenConfig
@@ -1,2 +1,0 @@
-dev.vality.codegen.generators.ValityErlangClientGenerator
-dev.vality.codegen.generators.ValityErlangServerGenerator
