+++ src/main/resources/META-INF/services/io.swagger.codegen.v3.CodegenConfig
@@ -1,0 +1,2 @@
+dev.vality.codegen.generators.v3.ValityErlangClientGenerator
+dev.vality.codegen.generators.v3.ValityErlangServerGenerator
